refactor(craft): extract sandbox manager factory from base.py
sandbox/base.py held get_sandbox_manager(), which lazy-imported the
concrete docker/k8s managers while both managers import base for the
ABC — a latent import cycle. Move the factory and its singleton to
sandbox/factory.py; base.py is now purely the abstract interface.
